Adjust padding utility classes
I've set the units to use rem instead of em, which will ensure that the space will remain consistent regardless of where it gets added. I also came across an issue where adding a horizontal padding class to an HTML element that has the container class would override that class' padding. In order to counter that, any horizontal padding is now assigned to a nested div instead.